Fix build and vendoring
The build was running glide install which installs the dependencies
found in the lock file. As a result we were running a very old version
of OPA.

These changes vendor the plugin dependencies and remove the glide step
from the build process. It's no longer required since all of the deps
are contained in the repo.

Also, remove broken plugin install from the README.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>