Use OPA SDK
Rewrite plug-in to use the new OPA SDK. This allows not only providing a policy
file from disk but rather a full OPA configuration, which in turn allows almost
all of the features known from using OPA as a standalone server, like bundles,
decision logging, and so on.

To preserve backwards compatibility, the old `-policy-file` is left pretty much
untouched. The single breaking change required for instead allowing the new
`-config-file` option was to remove the _default_ value for `-policy-file`, as
this would otherwise collide with `-config-file` as the options are mutually
exclusive.

Signed-off-by: Anders Eknert <anders@eknert.com>