chore: upgrade golangci-lint to v2.1.6 (#113)
Upgrades golangci-lint from v1.64.5 to v2.1.6.

This involves updating the version in the Makefile and
migrating the .golangci.yaml configuration to the new v2
format. This includes adjustments to linter and formatter
setups. A boolean expression was fixed by the linter in
main.go.

Previously the linter config had a typo, which caused
settings for 'lll' and 'misspell' to not get applied. This was
caused by a typo in the key name "linter-settings", which
should have been "linters-settings".

Signed-off-by: Ville Vesilehto <ville@vesilehto.fi>